+++ src/main/java/cn/sh1rocu/tacz/mixin/client/TextureAtlasMixin.java
@@ -1,6 +1,6 @@
 package cn.sh1rocu.tacz.mixin.client;
 
-import cn.sh1rocu.tacz.util.forge.ClientHooks;
+import cn.sh1rocu.tacz.api.event.TextureStitchEvent;
 import net.minecraft.client.renderer.texture.SpriteLoader;
 import net.minecraft.client.renderer.texture.TextureAtlas;
 import org.spongepowered.asm.mixin.Mixin;
@@ -12,6 +12,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class TextureAtlasMixin {
     @Inject(method = "upload", at = @At("TAIL"))
     private void tacz$uploadPost(SpriteLoader.Preparations preparations, CallbackInfo ci) {
-        ClientHooks.onTextureStitchedPost((TextureAtlas) (Object) this);
+        TextureStitchEvent.POST.invoker().post(new TextureStitchEvent.Post((TextureAtlas) (Object) this));
     }
 }