+++ src/main/java/com/tacz/guns/client/event/PlayerEnterWorld.java
@@ -1,6 +1,5 @@
 package com.tacz.guns.client.event;
 
-import cn.sh1rocu.tacz.api.event.PlayerEvent;
 import com.tacz.guns.resource.PackConvertor;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -9,25 +8,26 @@ import net.minecraft.network.chat.ClickEvent;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.HoverEvent;
 import net.minecraft.network.chat.Style;
+import net.minecraft.server.level.ServerPlayer;
 
 import java.io.File;
 
 @Environment(EnvType.CLIENT)
 public class PlayerEnterWorld {
-    public static void onPlayerEnterWorld(PlayerEvent.PlayerLoggedInEvent event) {
+    public static void onPlayerEnterWorld(ServerPlayer player) {
         File[] files = PackConvertor.FOLDER.toFile().listFiles();
         if (files != null && files.length > 0) {
-            event.getEntity().sendSystemMessage(pre(Component.translatable("message.tacz.convert_from_legacy.intro")));
-            event.getEntity().sendSystemMessage(pre(Component.translatable("message.tacz.convert_from_legacy.intro2")));
+            player.sendSystemMessage(pre(Component.translatable("message.tacz.convert_from_legacy.intro")));
+            player.sendSystemMessage(pre(Component.translatable("message.tacz.convert_from_legacy.intro2")));
             Component component = Component.translatable("message.tacz.convert_from_legacy")
                     .append(Component.translatable("message.tacz.convert_from_legacy.button")
                             .withStyle(Style.EMPTY.withColor(ChatFormatting.GREEN)
                                     .withClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/tacz convert"))
                                     .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.translatable("message.tacz.convert_from_legacy.hover"))
                                     )));
-            event.getEntity().sendSystemMessage(pre(component));
-            event.getEntity().sendSystemMessage(pre(Component.translatable("message.tacz.convert_from_legacy.hint")));
-            event.getEntity().sendSystemMessage(pre(Component.translatable("message.tacz.convert_from_legacy.hide")));
+            player.sendSystemMessage(pre(component));
+            player.sendSystemMessage(pre(Component.translatable("message.tacz.convert_from_legacy.hint")));
+            player.sendSystemMessage(pre(Component.translatable("message.tacz.convert_from_legacy.hide")));
         }
     }