+++ src/main/java/com/tacz/guns/event/HitboxHelperEvent.java
@@ -1,11 +1,11 @@
 package com.tacz.guns.event;
 
-import cn.sh1rocu.tacz.api.event.PlayerEvent;
 import cn.sh1rocu.tacz.api.event.PlayerTickEvent;
 import com.tacz.guns.config.common.OtherConfig;
 import com.tacz.guns.util.HitboxHelper;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.server.level.ServerPlayer;
 
 public class HitboxHelperEvent {
     public static void onPlayerTick(PlayerTickEvent.Post event) {
@@ -17,7 +17,7 @@ public class HitboxHelperEvent {
         }
     }
 
-    public static void onPlayerLoggedOut(PlayerEvent.PlayerLoggedOutEvent event) {
-        HitboxHelper.onPlayerLoggedOut(event.getEntity());
+    public static void onPlayerLoggedOut(ServerPlayer player) {
+        HitboxHelper.onPlayerLoggedOut(player);
     }
 }