+++ src/main/java/com/tacz/guns/init/CapabilityRegistry.java
@@ -1,8 +1,8 @@
 package com.tacz.guns.init;
 
-import cn.sh1rocu.tacz.api.event.EntityRemoveEvent;
 import com.tacz.guns.entity.sync.core.DataHolderCapabilityProvider;
 import com.tacz.guns.entity.sync.core.SyncedEntityData;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerEntityEvents;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
 import org.jetbrains.annotations.NotNull;
@@ -18,8 +18,7 @@ public class CapabilityRegistry implements EntityComponentInitializer {
     }
 
     public static void init() {
-        EntityRemoveEvent.EVENT.register(event -> {
-            var entity = event.getEntity();
+        ServerEntityEvents.ENTITY_UNLOAD.register((entity, level) -> {
             if (!(entity instanceof ServerPlayer)) {
                 DataHolderCapabilityProvider.CAPABILITY.maybeGet(entity).ifPresent(DataHolderCapabilityProvider::invalidate);
             }